feat: add assets to utils/go_to_page
